+++ src/main/java/com/ponomarevnikolaidiplom/services/classes/MedicalBillServiceImp.java
@@ -4,6 +4,8 @@ import com.ponomarevnikolaidiplom.dto.request.MedicalBillRequest;
 import com.ponomarevnikolaidiplom.dto.responce.MedicalBillResponce;
 import com.ponomarevnikolaidiplom.entities.MedicalBill;
 import com.ponomarevnikolaidiplom.entities.Specialization;
+import com.ponomarevnikolaidiplom.exceptionHandler.TypicalError;
+import com.ponomarevnikolaidiplom.exceptions.ServiceException;
 import com.ponomarevnikolaidiplom.repozitories.MedicalBillRepository;
 import com.ponomarevnikolaidiplom.repozitories.SpecializationRepository;
 import com.ponomarevnikolaidiplom.services.interfacies.MedicalBillService;
@@ -13,6 +15,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -29,7 +32,11 @@ public class MedicalBillServiceImp implements MedicalBillService {
     }
 
     @Override
-    public MedicalBillResponce getMedicalBill(Long id) {
+    public MedicalBillResponce getMedicalBill(Long id) throws ServiceException {
+        Optional<MedicalBill> request = medicalBillRepository.findById(id);
+        if (request.isEmpty()) {
+            throw new ServiceException("MedicalBill not found", TypicalError.NOT_FOUND);
+        }
         log.info("get MedicalBill by id={}", id);
         return convertMedicalBillToMedicalBillResponce(medicalBillRepository.findMedicalBillById(id));
     }
@@ -44,11 +51,16 @@ public class MedicalBillServiceImp implements MedicalBillService {
     }
 
     @Override
-    public String updateMedicalBill(MedicalBillRequest request) {
-        MedicalBill update = medicalBillRepository.getById(request.getId());
-        if (update == null) {
-            throw new RuntimeException("MedicalBill not found");
+    public String updateMedicalBill(MedicalBillRequest request) throws ServiceException {
+        if (request.getId() == null) {
+            throw new ServiceException("MedicalBill bad request id is null", TypicalError.BAD_REQUEST);
         }
+        Optional<MedicalBill> medicalBillOptional = medicalBillRepository.findById(request.getId());
+        if (medicalBillOptional.isEmpty()) {
+            throw new ServiceException("MedicalBill not found", TypicalError.NOT_FOUND);
+        }
+        MedicalBill update = medicalBillRepository.getById(request.getId());
+
         if (request.getName() != null) {
             update.setName(request.getName());
         }
@@ -62,44 +74,59 @@ public class MedicalBillServiceImp implements MedicalBillService {
     }
 
     @Override
-    public String deleteMedicalBill(Long id) {
+    public String deleteMedicalBill(Long id) throws ServiceException {
+        Optional<MedicalBill> medicalBillOptional = medicalBillRepository.findById(id);
+        if (medicalBillOptional.isEmpty()) {
+            throw new ServiceException("MedicalBill not found", TypicalError.NOT_FOUND);
+        }
         log.info("deleted MedicalBill id={}", id);
         medicalBillRepository.deleteById(id);
-        return "Услуга удалена id="+id;
+        return "deleted MedicalBill id=" + id;
     }
 
     @Override
-    public String addSpecializationToMedicalBill(Long idSpecialization, Long idMedicalBill) {
-        MedicalBill medicalBill=medicalBillRepository.findMedicalBillById(idMedicalBill);
-        Specialization specialization=specializationRepository.getById(idSpecialization);
-        if(medicalBill==null){
-            throw new RuntimeException("medicalBill not found");
+    public String addSpecializationToMedicalBill(Long idSpecialization, Long idMedicalBill) throws ServiceException {
+        Optional<MedicalBill> medicalBillOptional = medicalBillRepository.findById(idMedicalBill);
+
+        if (medicalBillOptional.isEmpty()) {
+            throw new ServiceException("MedicalBill not found", TypicalError.NOT_FOUND);
         }
-        if(specialization==null){
-            throw new RuntimeException("Specialization not found");
+
+        Optional<Specialization> specializationOptional = specializationRepository.findById(idSpecialization);
+
+        if (specializationOptional.isEmpty()) {
+            throw new ServiceException("Specialization not found", TypicalError.NOT_FOUND);
         }
+        MedicalBill medicalBill = medicalBillRepository.findMedicalBillById(idMedicalBill);
+        Specialization specialization = specializationRepository.getById(idSpecialization);
         medicalBill.getSpecializationList().add(specialization);
         specialization.getMedicalBillList().add(medicalBill);
         medicalBillRepository.saveAndFlush(medicalBill);
         specializationRepository.saveAndFlush(specialization);
-        return "Specialization name = "+specialization.getName()+" added to medicalBill name = "+medicalBill.getName();
+        return "Specialization name = " + specialization.getName() + " added to medicalBill name = " + medicalBill.getName();
     }
 
     @Override
-    public String deleteSpecializationFromMedicalBill(Long idSpecialization, Long idMedicalBill) {
-        MedicalBill medicalBill=medicalBillRepository.findMedicalBillById(idMedicalBill);
-        Specialization specialization=specializationRepository.getById(idSpecialization);
-        if(medicalBill==null){
-            throw new RuntimeException("Doctor not found");
+    public String deleteSpecializationFromMedicalBill(Long idSpecialization, Long idMedicalBill) throws ServiceException {
+        Optional<MedicalBill> medicalBillOptional = medicalBillRepository.findById(idMedicalBill);
+
+        if (medicalBillOptional.isEmpty()) {
+            throw new ServiceException("MedicalBill not found", TypicalError.NOT_FOUND);
         }
-        if(specialization==null){
-            throw new RuntimeException("Specialization not found");
+
+        Optional<Specialization> specializationOptional = specializationRepository.findById(idSpecialization);
+
+        if (specializationOptional.isEmpty()) {
+            throw new ServiceException("Specialization not found", TypicalError.NOT_FOUND);
         }
+
+        MedicalBill medicalBill = medicalBillRepository.findMedicalBillById(idMedicalBill);
+        Specialization specialization = specializationRepository.getById(idSpecialization);
         specialization.getMedicalBillList().remove(medicalBill);
         medicalBill.getSpecializationList().remove(specialization);
         medicalBillRepository.saveAndFlush(medicalBill);
         specializationRepository.saveAndFlush(specialization);
-        return "Specialization name = "+specialization.getName()+" deleted from MedicalBill name="+medicalBill.getName();
+        return "Specialization name = " + specialization.getName() + " deleted from MedicalBill name=" + medicalBill.getName();
     }
 
     private MedicalBill convertRequestToMedicalBill(MedicalBillRequest request) {
@@ -110,7 +137,7 @@ public class MedicalBillServiceImp implements MedicalBillService {
     private MedicalBillResponce convertMedicalBillToMedicalBillResponce(MedicalBill responce) {
         List<String> specializationList = new ArrayList<>();
         responce.getSpecializationList().forEach(specialization ->
-                specializationList.add("id="+specialization.getId()+", name="+specialization.getName()));
+                specializationList.add("id=" + specialization.getId() + ", name=" + specialization.getName()));
         return new MedicalBillResponce(responce.getId(),
                 responce.getName(),
                 responce.getPrice(),
